*: build phony etcd server binary for unsupported architectures

+++ storage/backend/backend.go
@@ -38,7 +38,7 @@ var (
 	// InitialMmapSize is the initial size of the mmapped region. Setting this larger than
 	// the potential max db size can prevent writer from blocking reader.
 	// This only works for linux.
-	InitialMmapSize = 10 * 1024 * 1024 * 1024
+	InitialMmapSize = int64(10 * 1024 * 1024 * 1024)
 )
 
 type Backend interface {
